+++ src/components/Captures/CaptureTable.test.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import {
   act,
   render,
@@ -13,10 +12,7 @@ import theme from '../common/theme';
 import { ThemeProvider } from '@material-ui/core/styles';
 import { SpeciesContext } from '../../context/SpeciesContext';
 import { TagsContext } from '../../context/TagsContext';
-import {
-  CapturesContext,
-  CapturesProvider,
-} from '../../context/CapturesContext';
+import { CapturesContext } from '../../context/CapturesContext';
 
 import CaptureTable from './CaptureTable';
 import FilterModel from '../../models/Filter';
@@ -40,6 +36,11 @@ describe('Captures', () => {
   // mock the treeTrackerApi
   const captureApi = require('../../api/treeTrackerApi').default;
 
+  captureApi.getCaptures = () => {
+    log.debug(`mock getCaptures: ${CAPTURES.captures}`);
+    return Promise.resolve(CAPTURES.captures);
+  };
+
   captureApi.getCaptureTags = () => {
     log.debug(`mock getCaptureTags: ${CAPTURE_TAGS}`);
     return Promise.resolve(CAPTURE_TAGS);
@@ -60,6 +61,8 @@ describe('Captures', () => {
       );
 
       render(component);
+      await act(() => captureApi.getCaptures());
+      await act(() => captureApi.getCaptureTags());
     });
 
     afterEach(cleanup);
@@ -76,77 +79,90 @@ describe('Captures', () => {
       expect(pageNums).toHaveLength(2);
     });
 
-    it('should show page # and capture count', () => {
+    it('should show page # and capture count', async () => {
       const counts = Array.from(
         document.querySelectorAll('.MuiTablePagination-caption')
       );
-      const arr = counts.map((count) => count.firstChild.textContent);
-      expect(arr[1]).toBe('1-4 of 4');
+      await waitFor(() => {
+        const arr = counts.map((count) => count.firstChild.textContent);
+        expect(arr[1]).toBe('1-4 of 4');
+      });
     });
 
-    it('should have 14 headers', () => {
-      const table = screen.getByRole(/table/i);
-      const headers = within(table).getAllByRole(/columnheader/i);
-      const arr = headers.map((header) => header.textContent);
-      expect(arr).toHaveLength(14);
+    it('should have 12 headers', async () => {
+      const table = await screen.findByRole(/table/i);
+      await waitFor(() => {
+        const headers = within(table).getAllByRole(/columnheader/i);
+        const arr = headers.map((header) => header.textContent);
+        expect(arr).toHaveLength(12);
+      });
     });
 
-    it('renders headers for captures table', () => {
-      const table = screen.getByRole(/table/i);
-      let item = screen.getAllByText(/Captures/i)[0];
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Capture ID/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Grower ID/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Device Identifier/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Planter Identifier/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Verification Status/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Species/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Token Id/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Wallet/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Capture Tags/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Created/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Latitude/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Longitude/i);
-      expect(item).toBeInTheDocument();
-      item = within(table).getByText(/Image URL/i);
-      expect(item).toBeInTheDocument();
+    it('renders headers for captures table', async () => {
+      const table = await screen.findByRole(/table/i);
+      await waitFor(() => {
+        // screen.logTestingPlaygroundURL(table);
+        // page header
+        let item = screen.getAllByText(/Captures/i)[0];
+        expect(item).toBeInTheDocument();
+        // table headers
+        item = within(table).getByText(/Capture ID/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Grower Acct\. ID/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Wallet/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Device Identifier/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Species/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Token Id/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Capture Tags/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Notes/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Created/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Longitude/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Latitude/i);
+        expect(item).toBeInTheDocument();
+        item = within(table).getByText(/Image URL/i);
+        expect(item).toBeInTheDocument();
+      });
     });
 
-    it('renders 4 rows ', () => {
-      const table = screen.getAllByRole(/rowgroup/i);
-      const rows = within(table[1]).getAllByRole(/row/i);
-      const arr = rows.map((link) => link.textContent);
-      expect(rows).toHaveLength(4);
+    it('renders 4 rows ', async () => {
+      const table = await screen.getAllByRole(/rowgroup/i);
+      await waitFor(() => {
+        const rows = within(table[1]).getAllByRole(/row/i);
+        const arr = rows.map((link) => link.textContent);
+        expect(rows).toHaveLength(4);
+      });
     });
 
-    it('renders links for planter ids (10-12)', () => {
-      const table = screen.getByTestId('captures-table-body');
-      const links = within(table).getAllByRole('link');
-      const arr = links.map((link) => link.textContent);
-      expect(arr.includes('10')).toBeTruthy();
-      expect(arr.includes('11')).toBeTruthy();
-      expect(arr.includes('12')).toBeTruthy();
+    it('renders links for planter ids (10-12)', async () => {
+      const table = await screen.getByTestId('captures-table-body');
+      await waitFor(() => {
+        const links = within(table).getAllByRole('link');
+        const arr = links.map((link) => link.textContent);
+        expect(arr.includes('10')).toBeTruthy();
+        expect(arr.includes('11')).toBeTruthy();
+        expect(arr.includes('12')).toBeTruthy();
+      });
     });
 
-    it('displays captures data', () => {
-      const table = screen.getByTestId('captures-table-body');
-      const status = within(table).getAllByText(/approved/i);
-      expect(status).toHaveLength(2);
-      const device = within(table).getAllByText(/1-abcdef123456/i);
-      expect(device).toHaveLength(1);
-      const captureTag = within(table).getAllByText(/tag_c/i);
-      expect(captureTag).toHaveLength(4);
+    it('displays captures data', async () => {
+      const table = await screen.getByTestId('captures-table-body');
+      await waitFor(() => {
+        const wallet = within(table).getAllByText(/grower3@some.place/i);
+        expect(wallet).toHaveLength(2);
+        const device = within(table).getAllByText(/1-abcdef123456/i);
+        expect(device).toHaveLength(1);
+        // const captureTag = within(table).getAllByText(/tag_c/i);
+        // expect(captureTag).toHaveLength(4);
+      });
     });
   });
 
@@ -165,8 +181,8 @@ describe('Captures', () => {
       allIds: [],
       byId: {},
       filter: new FilterModel(),
-      // queryCapturesApi: jest.fn(),
-      queryCapturesApi: () => {},
+      // getCaptures: jest.fn(),
+      getCaptures: () => {},
       getCaptureCount: () => {},
       getCapturesAsync: () => {},
     };
@@ -181,25 +197,27 @@ describe('Captures', () => {
     //   .mockReturnValueOnce({ data });
 
     // PASSES TESTS BUT STILL DOESN'T RETURN DATA
-    context.queryCapturesApi = jest.fn(() => {
-      console.log('mock queryCapturesApi');
+    const captureApi = require('../../api/treeTrackerApi').default;
+
+    captureApi.getCaptures = jest.fn(() => {
+      log.debug('mock getCaptures');
       // return Promise.resolve({ data });
       return axios.get
         .mockReturnValueOnce({
-          data: { count: data.length },
+          data: { count: data.length, captures: CAPTURES.captures },
         })
         .mockReturnValueOnce({ data });
     });
 
-    context.queryCapturesApi = jest.fn(() => {
-      console.log('mock queryCapturesApi');
-      // return Promise.resolve({ data });
-      return axios.get
-        .mockReturnValueOnce({
-          data: { count: data.length },
-        })
-        .mockReturnValueOnce({ data });
-    });
+    // context.getCaptures = jest.fn(() => {
+    //  log.debug('mock getCaptures');
+    //   // return Promise.resolve({ data });
+    //   return axios.get
+    //     .mockReturnValueOnce({
+    //       data: { count: data.length },
+    //     })
+    //     .mockReturnValueOnce({ data });
+    // });
 
     beforeEach(async () => {
       component = (
@@ -212,40 +230,26 @@ describe('Captures', () => {
 
       render(component);
       act(() => axios.get());
-      act(() => context.queryCapturesApi()); //passes tests
-      // await act(async () => await context.queryCapturesApi()); //stops data errors
+      act(() => captureApi.getCaptures());
     });
 
     afterEach(cleanup);
 
     it('should make a request for capture count', () => {
       expect(axios.get).toHaveBeenCalled();
-      // console.log(axios.get.mock.calls);
+      //log.debug(axios.get.mock.calls);
       expect(axios.get.mock.calls[1][0]).toContain(`trees/count?`);
     });
 
     it('should call captures API with a valid filter', () => {
       const filter = JSON.stringify({
-        where: { approved: true, active: true },
-        order: ['id asc'],
+        status: 'approved',
         limit: 25,
-        skip: 0,
-        fields: {
-          id: true,
-          timeCreated: true,
-          status: true,
-          active: true,
-          approved: true,
-          planterId: true,
-          planterIdentifier: true,
-          deviceIdentifier: true,
-          speciesId: true,
-          tokenId: true,
-        },
+        offset: 0,
       });
 
       expect(axios.get).toHaveBeenCalled();
-      expect(axios.get.mock.calls[0][0]).toContain(`/trees?filter=${filter}`);
+      expect(axios.get.mock.calls[0][0]).toContain(`/v2/captures`);
     });
   });
 });