+++ src/components/Growers/Grower.js
@@ -8,7 +8,7 @@ import {
   CardContent,
 } from '@material-ui/core';
 import { Person } from '@material-ui/icons';
-import LinkToWebmap from '../common/LinkToWebmap';
+import LinkToWebmap, { pathType } from '../common/LinkToWebmap';
 import OptimizedImage from '../OptimizedImage';
 import GrowerOrganization from 'components/GrowerOrganization';
 import { useStyle, GROWER_IMAGE_SIZE } from './Growers.styles.js';
@@ -52,7 +52,11 @@ export const Grower = ({ grower, growerClick }) => {
               {grower.firstName} {grower.lastName}
             </Typography>
             <Typography>
-              ID: <LinkToWebmap value={grower.id} type="user" />
+              ID:{' '}
+              <LinkToWebmap
+                value={grower.reference_id || grower.id}
+                type={pathType.planter}
+              />
             </Typography>
             <GrowerOrganization
               organizationName={grower?.organization}