+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## [1.120.2](https://github.com/Greenstand/treetracker-admin-client/compare/v1.120.1...v1.120.2) (2023-01-01)
+
+
+### Bug Fixes
+
+* used the correct property for verification status ([#996](https://github.com/Greenstand/treetracker-admin-client/issues/996)) ([120e50a](https://github.com/Greenstand/treetracker-admin-client/commit/120e50a01f3bb86082b4c5d1ddd37726cf47d2d4))
+
+## [1.120.1](https://github.com/Greenstand/treetracker-admin-client/compare/v1.120.0...v1.120.1) (2022-12-31)
+
+
+### Bug Fixes
+
+* change web map links ([#973](https://github.com/Greenstand/treetracker-admin-client/issues/973)) ([6494e52](https://github.com/Greenstand/treetracker-admin-client/commit/6494e52e7b64cef5904d172a9153972a0b0f8dbb))
+
+# [1.120.0](https://github.com/Greenstand/treetracker-admin-client/compare/v1.119.1...v1.120.0) (2022-12-30)
+
+
+### Features
+
+* update captures to use treetracker microservice ([f1164f6](https://github.com/Greenstand/treetracker-admin-client/commit/f1164f673fde148357f69e1c718a8367542f8eb9))
+* update filters to use keys that match new apis ([2f9ba61](https://github.com/Greenstand/treetracker-admin-client/commit/2f9ba61b66e841b0241d4401bf6352747065a71a))
+
+## [1.119.1](https://github.com/Greenstand/treetracker-admin-client/compare/v1.119.0...v1.119.1) (2022-12-28)
+
+
+### Bug Fixes
+
+* filtering for "All" values in Earnings Table filter ([#989](https://github.com/Greenstand/treetracker-admin-client/issues/989)) ([d0f0f81](https://github.com/Greenstand/treetracker-admin-client/commit/d0f0f8130dd362a000d8c23aaeb6e3903babe046))
+
 # [1.119.0](https://github.com/Greenstand/treetracker-admin-client/compare/v1.118.0...v1.119.0) (2022-12-19)